refactor : All Package and Class Infrastructure Has Been Refactored and Unused Classes Have Been Deleted

+++ src/main/java/org/gelecekbilimde/scienceplatform/media/controller/MediaController.java
@@ -1,8 +1,8 @@
 package org.gelecekbilimde.scienceplatform.media.controller;
 
 import lombok.RequiredArgsConstructor;
-import org.gelecekbilimde.scienceplatform.common.Response;
-import org.gelecekbilimde.scienceplatform.media.enums.MediaContentType;
+import org.gelecekbilimde.scienceplatform.common.model.response.Response;
+import org.gelecekbilimde.scienceplatform.media.model.enums.MediaContentType;
 import org.gelecekbilimde.scienceplatform.media.service.MediaService;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -23,10 +23,11 @@ class MediaController {
 	@PostMapping
 	@PreAuthorize("hasAuthority('post:create')")
 	public Response<List<Object>> secretVersionRole(@RequestParam("files") List<MultipartFile> files,
-													@RequestParam("groupId") Integer groupId,
+													@RequestParam("groupId") Long groupId,
 													@RequestParam("contentType") MediaContentType contentType) {
 
 		//todo size kontrol
 		return Response.ok(mediaService.uploadMedia(groupId, contentType, files));
 	}
+
 }
